+++ modules/gui/builtin/cpp/Gateway.cpp
@@ -38,7 +38,7 @@ static const nlsGateway gateway[] = {
     { "banner", (void*)Nelson::GuiGateway::bannerBuiltin, 0, 0, CPP_BUILTIN_WITH_EVALUATOR },
     { "inserthtml", (void*)Nelson::GuiGateway::inserthtmlBuiltin, 0, 1,
         CPP_BUILTIN_WITH_EVALUATOR },
-    { "lookandfeel", (void*)Nelson::GuiGateway::lookandfeelBuiltin, 1, 2},
+    { "lookandfeel", (void*)Nelson::GuiGateway::lookandfeelBuiltin, 1, 2 },
     { "uigetdir", (void*)Nelson::GuiGateway::uigetdirBuiltin, 1, 2 },
     { "qt_verbose", (void*)Nelson::GuiGateway::qt_verboseBuiltin, 1, 1 },
 };